requirements.txt
streamlit
scipy.stats
numpy
